plant_disease_rag/requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
openai